Postprocessing order (#3)
* add attributes , ,  etc to sol in postprocessing in equations.py

* add test for creation of ,  etc from previous commit

* move postprocessing of inflation start and end from to inflation.py

`postprocessing_inflation_start` and `postprocessing_inflation_end` were
previously in time/inflation.py and are now moved to inflation.py. This
is now possible because we can use `N_events` to determine `N_beg` and
`N_end` and don't have to take the detour via `t_events`. This will make
the addition of other independent variables easier since those functions
won't need to be written.

The postprocessing order is now:
1) Equations.sol(sol) in equations.py
2) InflationEquations.sol(sol) in inflation.py
3) InflationEquationsT.sol(sol) in time/inflation.py

* adapt tests to reflect changes from previous commit

* add tests for basic methods in InflationEquations and InflationEquationsT including tests for errors

Co-authored-by: lukas hergt <lh561@cam.ac.uk>