use RK45 instead of DOP853 after all
The guideline for using DOP853 over RK45 when tolerances are low is because
DOP853 uses fewer function evaluations (steps) in that case, thus it is quicker
for low tolerances (but not more accurate, accuracy is determined by the
tolerances). For Oscode fewer steps is actually not desirable and the
difference in speed of the background integration is negligible compared to
Oscode's own speed. Thus, RK45 is probably actually more helpful here.