add convenience functions get_w_reh and conformal_time_ratio
get_w_reh determines the equation of state parameter from the comoving Hubble
horizon with respect to e-folds.

conformal_time_ratio determines the ratio of conformal time during and before
inflation to after the end of inflation until today.

Both these functions help as parameter constraints on sets that violate
reheating conditions or the solving of the horizon problem.