use exp(-2N) instead of 1/a and pass the end index to Perturbation class
For large amounts of inflation (large N), computing the scale factor a can
cause a stackoverflow.  Replace all instances of using a directly with an
appropriate use of N instead.  This stackoverflow error is particularly
problematic towards the end of the arrays. To avoid this in the frequency and
damping calculations, limit those arrays to the end of the perturbation
integration which is calculated anyhow.